+++ InvestorsAssist/InvestorsAssist.Entities/Company.cs
@@ -11,7 +11,7 @@ namespace InvestorsAssist.Entities
         public string Symbol { get; set; }
         public string Exchange { get; set; }
         public string Name { get; set; }
-        public float LastSale { get; set; }
+        public double LastSale { get; set; }
         public decimal MarketCap { get; set; }
         public string Sector { get; set; }
         public string Industry { get; set; }
